extract spec to the own one
